REFACTOR: Extract builder components and reorganize logic folder
