+++ src/logic/builders/GripperBuilder.ts
@@ -1,0 +1,8 @@
+import { ToolBuilder } from './ToolBuilder'
+import Gripper from '@/model/tools/Gripper'
+
+export class GripperBuilder implements ToolBuilder<Gripper> {
+  build() {
+    return new Gripper()
+  }
+}
