+++ src/logic/builders/ToolBuilder.ts
@@ -1,0 +1,3 @@
+export interface ToolBuilder<T> {
+  build(): T
+}
