+++ src/logic/interpreter/Lexer.ts
